Exceptions move part4
Problem:
Currently exception classes are scattered all over SDK, this will lead to unecessary duplication and confusion

Analysis:
Corrected UnsupportedMethod imports

Tests:
Flake8
