+++ f5/bigip/tm/sys/test/unit/test_httpd.py
@@ -16,8 +16,8 @@
 import mock
 import pytest
 
-from f5.bigip.mixins import UnsupportedMethod
 from f5.bigip.tm.sys.httpd import Httpd
+from f5.sdk_exception import UnsupportedMethod
 
 
 @pytest.fixture
