+++ f5/bigip/tm/sys/test/unit/test_performance.py
@@ -16,9 +16,9 @@
 import mock
 import pytest
 
-from f5.bigip.mixins import UnsupportedMethod
 from f5.bigip.resource import UnsupportedOperation
 from f5.bigip.tm.sys.performance import Performances
+from f5.sdk_exception import UnsupportedMethod
 
 
 @pytest.fixture
