+++ f5/bigiq/tm/sys/test/unit/test_dns.py
@@ -16,8 +16,8 @@
 import mock
 import pytest
 
-from f5.bigip.mixins import UnsupportedMethod
 from f5.bigip.tm.sys.dns import Dns
+from f5.sdk_exception import UnsupportedMethod
 
 
 @pytest.fixture
